Make auth token work.
