+++ packages/mcp/package.json
@@ -91,6 +91,7 @@
     "@breadboard-ai/types": "^0.8.0",
     "@breadboard-ai/utils": "^0.0.2",
     "@modelcontextprotocol/sdk": "^1.17.1",
+    "idb": "^8.0.3",
     "zod": "^3.23.8"
   }
 }
